requirements.txt
-e .

Django==1.8.4
Pillow==2.9.0
djangorestframework==3.2.3
sorl-thumbnail==12.3
wsgiref==0.1.2

# Test
coverage
flake8
tox
mock

# Build
pypandoc
wheel
